Make byte-compile-cond-use-jump-table obsolete
This user option was introduced in 2017, with the rationale:

    "This is a workaround for when `byte-compile-cond-jump-table'
    accidentally generates wrong code (hasn't happened so far in my
    tests), and should be removed once we're sure there are no issues
    with it."  https://lists.gnu.org/r/emacs-devel/2017-02/msg00223.html

There hasn't been any bugs in this area within the last 6 years, and the
few we did see before that were quickly resolved.  Let's declare this
variable obsolete now, while keeping its behavior when set to nil.  We
can always revert the obsoletion if it turns out to be needed after all.

* lisp/emacs-lisp/bytecomp.el
(byte-compile-cond-use-jump-table): Declare obsolete.