[MPS] Scan thread->m_getcjmp ambigously
This makes it necessary to allocate the buffer indirectly, as it needs
to be an MPS root rather than part of a normal pseudovector.

* src/thread.c (main_thread_getcjmp): New variable.
(main_thread): Use it.
(run_thread): Free setjmp buffer when tearing down a thread.
(Fmake_thread): Allocate setjmp buffer when starting a new thread.
* src/thread.h (m_getcjmp): Point to sys_jmp_buf indirectly.
* src/igc.c (root_create_main_thread): Create an ambiguous root for
'main_thread_getjmp'.