(register-preview-info): Delete slot `smatch`
Use the `types` slot to carry that info instead.

Replace the list of types `(all)` with `(t)` since `t` is the
usual name of the "supertype of all types".
Use the type `null` to represent the fact that empty registers
can be used.  Allow an empty list of types to stand for
`(t null)`, i.e. any register even empty ones.

* lisp/register.el (register-preview-info): Delete slot `smatch`.
(register-command-info): Delete arg `:smatch`.  Adjust `:types` instead.
Fix the `copy-rectangle-to-register` case, which disallowed using
an empty register.
(register-of-type-alist): Adjust handling of `types` accordingly.
(register-preview-1): Simplify.
(register-read-with-preview-fancy): Use types instead of `smatch`.
Use it also in place of `act`.