(register-read-with-preview-fancy): Fix handling of control chars
The code assumed that the string returned by `register-preview-function`
has the register name as the first char.  This was an
incompatible change which broke packages that set this var,
such as `calculator.el` and others.
Remove this assumption by recording the register names in the
preview buffer on a new `register--name` text property.
While at it, fix a few other problems where control chars were
not pretty printed.

* lisp/register.el (register-preview-1): Remember the raw
register name in the `register--name` text property.
(register-preview-forward-line): Use the
`register--name` text property.
(register--find-preview): New function.
(register-read-with-preview-fancy): Use it.
If the last command inserted more than one char, only keep the first of
the new chars.
Make sure control chars are pretty printed in the minibuffer.
including minibuffer messages.